feat: add dashboard screen data fetch
- Add `DashboardViewModel` to fetch device data.
- Create `DeviceDto`, `ConfigurationDto`,
 `FeatureDto`, `MeasureTypeDto`, `CapacityDto` and rename `User` to `UserDto`.
- Update `ComwattApi` to fetch devices.
- Add test for `ComwattApi` to check fetch devices.
- Update `strings.xml` with new texts.
- Update `HomeScreen` to display device data.

+++ composeApp/src/commonMain/kotlin/net/thevenot/comwatt/ui/home/HomeScreenState.kt
@@ -13,6 +13,6 @@ data class HomeScreenState(
     val injectionGaugeEnabled: Boolean = true,
     val withdrawalsGaugeEnabled: Boolean = true,
     val lastRefreshInstant: Instant? = null,
-    val timeDifference: Long? = null,
+    val timeDifference: Int? = null,
     val siteTimeSeries: SiteTimeSeries = SiteTimeSeries()
 )
